Significant Memory Improvements.
Through careful pre-allocation, managed to remove the huge 30GiB memory estimate for the a_grid_calc() function. This led to a significant speed up (at least half the previous total time). Through this method, it was necessary/useful to be as specific as possible with the types of function arguments. Within the most inner function (len_calc()), it was found that the indexing method, calculating idx, was the cause of the large memory and so the simpler method of an if statement within a for loop over all faces was employed.