Fixing Culling Mistake.
Had previously culled all back-facing triangles. This is usual in ray-tracing but since we start in the sample here, we need to cull all front-facing triangles. Fixed this error in our choice of valid faces for all notebooks. Also, in the process of investigating duplicate path length problems in multiple crystal program and no path length outputs in all programs. These program versions are still not complete.