Completion of Sampling Function
Completed the function (within MC_Sampling.ipynb) that will calculate the required number of coordinates within the sample. Updated AttenuationCorrection.jl with the new method and with the ability to perform single crystal and more complex general morphologies. This led to some errors which are yet to be solved. Am yet to update the sampling method in (SingleCrystal/General)AttenTest2.ipynb