Flux+Absorption Correction Function
Created the main (uppermost) function to correct the set of inputted nxspe files for flux and absorption. Located in DataCorrection.jl which replaced AttenuationCorrection.jl . Removed the use of sparse matrices. They had a minor advantage time and memory wise but removed information as to location of NaN versus 0 which was necessary to output the corrected data.